requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.13.1=pypi_0
aiosignal=1.4.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
appdirs=1.4.4=pypi_0
attrs=25.4.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
blis=1.3.0=pypi_0
bs4=0.0.2=pypi_0
bzip2=1.0.8=h0ad9c76_8
ca-certificates=2025.10.5=h4c7d964_0
catalogue=2.0.10=pypi_0
certifi=2025.6.15=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.3.0=pypi_0
cloudpathlib=0.23.0=pypi_0
colorama=0.4.6=pypi_0
confection=0.1.5=pypi_0
cymem=2.0.11=pypi_0
dataclasses-json=0.6.7=pypi_0
datasets=4.3.0=pypi_0
dill=0.4.0=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
docstring-parser=0.17.0=pypi_0
filelock=3.20.0=pypi_0
frozenlist=1.8.0=pypi_0
fsspec=2025.9.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
greenlet=3.2.4=pypi_0
h11=0.16.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
httpx-sse=0.4.3=pypi_0
huggingface-hub=0.36.0=pypi_0
idna=3.10=pypi_0
instructor=1.11.3=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.10.0=pypi_0
joblib=1.5.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
langchain=1.0.1=pypi_0
langchain-classic=1.0.0=pypi_0
langchain-community=0.4=pypi_0
langchain-core=1.0.0=pypi_0
langchain-openai=1.0.0=pypi_0
langchain-text-splitters=1.0.0=pypi_0
langcodes=3.5.0=pypi_0
langgraph=1.0.1=pypi_0
langgraph-checkpoint=3.0.0=pypi_0
langgraph-prebuilt=1.0.1=pypi_0
langgraph-sdk=0.2.9=pypi_0
langsmith=0.4.37=pypi_0
language-data=1.3.0=pypi_0
libexpat=2.7.1=hac47afa_0
libffi=3.5.2=h52bdfb6_0
liblzma=5.8.1=h2466b09_2
liblzma-devel=5.8.1=h2466b09_2
libmpdec=4.0.0=h2466b09_0
libsqlite=3.50.4=hf5d6505_0
libzlib=1.3.1=h2466b09_2
marisa-trie=1.3.1=pypi_0
markdown=3.8.2=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=pypi_0
marshmallow=3.26.1=pypi_0
mdurl=0.1.2=pypi_0
multidict=6.7.0=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.13=pypi_0
mypy-extensions=1.1.0=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.5=pypi_0
numpy=2.3.0=pypi_0
openai=1.109.1=pypi_0
openssl=3.5.4=h725018a_0
orjson=3.11.3=pypi_0
ormsgpack=1.11.0=pypi_0
packaging=25.0=pypi_0
pandas=2.3.3=pypi_0
pdf2image=1.17.0=pypi_0
pillow=12.0.0=pypi_0
pip=25.2=pyh145f28c_0
preshed=3.0.10=pypi_0
propcache=0.4.1=pypi_0
pyarrow=21.0.0=pypi_0
pycryptodome=3.23.0=pypi_0
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pydantic-settings=2.11.0=pypi_0
pygments=2.19.2=pypi_0
pymupdf=1.26.5=pypi_0
pypdf2=3.0.1=pypi_0
pytesseract=0.3.13=pypi_0
python=3.13.0=hf5aa216_101_cp313
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.0=pypi_0
python_abi=3.13=8_cp313
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
ragas=0.3.7=pypi_0
regex=2024.11.6=pypi_0
reportlab=4.4.4=pypi_0
requests=2.32.5=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=14.2.0=pypi_0
scikit-learn=1.7.2=pypi_0
scikit-network=0.33.3=pypi_0
scipy=1.16.2=pypi_0
setuptools=80.9.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smart-open=7.4.4=pypi_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.7=pypi_0
spacy=3.8.7=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
spacy-pkuseg=1.0.1=pypi_0
sqlalchemy=2.0.44=pypi_0
srsly=2.5.1=pypi_0
tenacity=9.1.2=pypi_0
thinc=8.3.6=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.9.0=pypi_0
tk=8.6.13=h2c6b04d_2
tqdm=4.67.1=pypi_0
typer=0.20.0=pypi_0
typing-extensions=4.14.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025.2=pypi_0
ucrt=10.0.26100.0=h57928b3_0
urllib3=2.5.0=pypi_0
vc=14.3=h2b53caa_32
vc14_runtime=14.44.35208=h818238b_32
vcomp14=14.44.35208=h818238b_32
wasabi=1.1.3=pypi_0
weasel=0.4.1=pypi_0
wrapt=2.0.0=pypi_0
xxhash=3.6.0=pypi_0
xz=5.8.1=h208afaa_2
xz-tools=5.8.1=h2466b09_2
yarl=1.22.0=pypi_0
zh-core-web-sm=3.8.0=pypi_0
zstandard=0.25.0=pypi_0
